+++ client/app/form-builder/adv-form-builder-def.js
@@ -63,6 +63,36 @@ var advFormBuilderDef = {
       "linkId": "/_externallyDefined"
     },
     {
+      "questionCode": "_linkId",
+      "question": "You should never see this!",
+      "dataType": "ST",
+      "header": false,
+      "displayControl": {
+        "css": [
+          {
+            "name": "display",
+            "value": "none"
+          }
+        ]
+      },
+      "linkId": "/_linkId"
+    },
+    {
+      "questionCode": "_questionCode",
+      "question": "You should never see this!",
+      "dataType": "ST",
+      "header": false,
+      "displayControl": {
+        "css": [
+          {
+            "name": "display",
+            "value": "none"
+          }
+        ]
+      },
+      "linkId": "/_questionCode"
+    },
+    {
       "questionCode": "__itemTypeRef",
       "question": "You should never see this!",
       "dataType": "ST",
@@ -160,13 +190,20 @@ var advFormBuilderDef = {
       "skipLogic": {
         "conditions": [
           {
-            "source": "/_isHeader",
+            "source": "/__itemTypeRef",
             "trigger": {
-              "value": "No"
+              "value": "question"
+            }
+          },
+          {
+            "source": "/_dataType",
+            "trigger": {
+              "notEqual": "__CNE_OR_CWE__"
             }
           }
         ],
-        "action": "show"
+        "action": "show",
+        "logic": "ALL"
       },
       "items": [
         {
@@ -1084,12 +1121,127 @@ var advFormBuilderDef = {
       "linkId": "/copyrightNotice"
     },
     {
+      "questionCode": "_observationLinkPeriod",
+      "question": "Add link to FHIR Observations?",
+      "dataType": "CNE",
+      "answers": "boolean",
+      "value": {
+        "text": "No",
+        "code": false
+      },
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "/__itemTypeRef",
+            "trigger": {
+              "notEqual": "display"
+            }
+          }
+        ],
+        "action": "show"
+      },
+      "items": [
+        {
+          "questionCode": "duration",
+          "question": "Time window*",
+          "required": true,
+          "dataType": "INT",
+          "linkId": "/_observationLinkPeriod/duration",
+          "skipLogic": {
+            "conditions": [
+              {
+                "source": "/_observationLinkPeriod",
+                "trigger": {
+                  "value": {"code": true}
+                }
+              },
+              {
+                "source": "/_questionCode",
+                "trigger": {
+                  "exists": true
+                }
+              }
+            ],
+            "action": "show",
+            "logic": "ALL"
+          }
+        },
+        {
+          "questionCode": "unit",
+          "question": "Unit*",
+          "required": true,
+          "dataType": "CNE",
+          "answers": "observationLinkPeriodUnits",
+          "displayControl": {
+            "answerLayout": {
+              "type": "COMBO_BOX"
+            }
+          },
+          "linkId": "/_observationLinkPeriod/unit",
+          "skipLogic": {
+            "conditions": [
+              {
+                "source": "/_observationLinkPeriod",
+                "trigger": {
+                  "value": {"code": true}
+                }
+              },
+              {
+                "source": "/_questionCode",
+                "trigger": {
+                  "exists": true
+                }
+              }
+            ],
+            "action": "show",
+            "logic": "ALL"
+          }
+        },
+        {
+          "questionCode": "linkIdWarning",
+          "question": "Please enter a code for the question in the Basic Properties tab.",
+          "dataType": "TITLE",
+          "linkId": "/_observationLinkPeriod/linkIdWarning",
+          "skipLogic": {
+            "conditions": [
+              {
+                "source": "/_observationLinkPeriod",
+                "trigger": {
+                  "value": {"code": true}
+                }
+              },
+              {
+                "source": "/_questionCode",
+                "trigger": {
+                  "exists": false
+                }
+              }
+            ],
+            "action": "show",
+            "logic": "ALL"
+          }
+        }
+      ],
+      "linkId": "/_observationLinkPeriod"
+    },
+    {
       "questionCode": "_fhirVariables",
       "question": "FHIR Variable",
       "header": true,
       "questionCardinality": {
         "min": "1",
         "max": "*"
+      },
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "/__itemTypeRef",
+            "trigger": {
+              "notEqual": "display"
+            }
+          }
+        ],
+        "action": "show"
       },
       "items": [
         {