+++ client/app/form-builder/form-builder-sidenav.controller.js
@@ -749,6 +749,8 @@
                   if(newValue !== oldValue) {
                     scope.selectedNode.isDirty = true;
                     scope.selectedNode.skipLogicDirty = true;
+                    scope.selectedNode.lfData.advanced.itemHash['/_questionCode/1'].value =
+                        newValue && newValue.length > 0 ? newValue : null;
                   }
                 }
               }, true);
@@ -760,6 +762,8 @@
                   if(newValue !== oldValue) {
                     scope.selectedNode.isDirty = true;
                     scope.selectedNode.skipLogicDirty = true;
+                    scope.selectedNode.lfData.advanced.itemHash['/_linkId/1'].value =
+                        newValue && newValue.length > 0 ? newValue : null;
                   }
                 }
               }, true);