+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [6.1.0]  2020-09-08
+### Added
+- Added support for observation link period.
+### Fixed
+- Fixed a bug in skip logic, when skip logic source is pointing to its parent which is a question type.
+
 ## [6.0.1]  2020-08-05
 ### Fixed
 - Fixed missing node id labels in the auto complete box for skip logic sources.